3.2_crud/stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
django-filter==23.2
